show only user workspaces list
