+++ applications/fishing-map/src/features/app/App.tsx
@@ -18,7 +18,8 @@ import Map from 'features/map/Map'
 import MapWorkspacesList from 'features/map-workspaces-list/MapWorkspacesList'
 import Timebar from 'features/timebar/Timebar'
 import Sidebar from 'features/sidebar/Sidebar'
-import { isUserAuthorized, isUserLogged, logoutUserThunk } from 'features/user/user.slice'
+import { logoutUserThunk } from 'features/user/user.slice'
+import { isUserAuthorized, isUserLogged } from 'features/user/user.selectors'
 import { HOME, WORKSPACE } from 'routes/routes'
 import { updateLocation } from 'routes/routes.actions'
 import { selectSidebarOpen } from './app.selectors'
